+++ test/middleware.js
@@ -106,12 +106,46 @@ describe('middleware', () => {
     let server = app.listen();
 
     // Test
-    const response = await request(server).get('/').set('X-Request-ID', 'test-value');
+    const response = await request(server)
+      .get('/')
+      .set('X-Request-ID', 'test-value');
     expect(response.status).toEqual(200);
     expect(response.type).toEqual('application/json');
     expect(response.body.data).toEqual('Hello World!');
     expect(response.headers['x-request-id']).toBeDefined();
     expect(response.headers['x-request-id']).toEqual('test-value');
+
+    // Teardown
+    server.close();
+    done();
+  });
+
+  test('Should use custom uuid function if provided, and return the correct value', async done => {
+    // Setup
+    const app = new Koa();
+    const logger = new KoaReqLogger({
+      uuidFunction: () => {
+        return 'test-uuid';
+      }
+    });
+    app.use(logger.getMiddleware());
+
+    app.use((ctx, next) => {
+      ctx.status = 200;
+      ctx.body = {
+        data: 'Hello World!'
+      };
+    });
+
+    let server = app.listen();
+
+    // Test
+    const response = await request(server).get('/');
+    expect(response.status).toEqual(200);
+    expect(response.type).toEqual('application/json');
+    expect(response.body.data).toEqual('Hello World!');
+    expect(response.headers['x-request-id']).toBeDefined();
+    expect(response.headers['x-request-id']).toEqual('test-uuid');
 
     // Teardown
     server.close();